Tidied codebase slightly, fixed release mode bug
Validation layers were complaining whilst running in release mode because pQueueFamilyIndices member of VkSwapchainCreateInfoKHR object was a dangling pointer that went out of scope after returning from init function

+++ Cassidy/Utils/GlobalTimer.cpp
@@ -7,6 +7,6 @@ void GlobalTimer::updateGlobalTimerImpl()
   uint64_t currentTimeMS = SDL_GetTicks64();
   const double currentTimeSecs = currentTimeMS * 0.001;
 
-  m_deltaTimeSecs = currentTimeSecs - m_engineTimeSecs;
+  m_deltaTimeSecs = static_cast<float>(currentTimeSecs - m_engineTimeSecs);
   m_engineTimeSecs = currentTimeSecs;
 }
